no build warnings

+++ pe32.cpp
@@ -12,10 +12,16 @@
 #include <set>
 #include <boost/foreach.hpp>
 
-typedef std::map<int, bool> digit_map;
-bool is_pandigital_for_range(int start, int end, size_t* ints, int ints_size);
-bool spy_the_digits(int min, int max, size_t n, digit_map& reqd_digits);
-int len_n(size_t min);
+typedef std::map<size_t, bool> digit_map;
+bool is_pandigital_for_range(size_t start,
+                             size_t end,
+                             size_t* ints,
+                             size_t ints_size);
+bool spy_the_digits(size_t min,
+                    size_t max,
+                    size_t n,
+                    digit_map& reqd_digits);
+size_t len_n(size_t min);
 
 std::string& pe32::name()
 {
@@ -43,15 +49,15 @@ void pe32::run()
      *
      *
      */
-    int total = 0;
-    std::set<int> ok_prods;
-    for(int mand = 1; mand < 10000; ++mand)
+    size_t total = 0;
+    std::set<size_t> ok_prods;
+    for(size_t mand = 1; mand < 10000; ++mand)
     {
-        for(int mier = 1; mier < 10000; ++mier)
+        for(size_t mier = 1; mier < 10000; ++mier)
         {
-            int prod = mand * mier;
+            size_t prod = mand * mier;
             size_t nums[] = { mand, mier, prod };
-            int full_len = len_n(mand);
+            size_t full_len = len_n(mand);
             full_len += len_n(mier);
             full_len += len_n(prod);
             if ( full_len != 9 ) { continue; }
@@ -63,7 +69,7 @@ void pe32::run()
     }
 
     // sum the prods
-    BOOST_FOREACH ( int prod, ok_prods )
+    BOOST_FOREACH ( size_t prod, ok_prods )
     {
         total += prod;
     }
@@ -71,12 +77,15 @@ void pe32::run()
     std::cout << "PE32 " << total << std::endl;
 }
 
-bool is_pandigital_for_range(int start, int end, size_t* ints, int ints_size)
+bool is_pandigital_for_range(size_t start,
+                             size_t end,
+                             size_t* ints,
+                             size_t ints_size)
 {
     digit_map reqd_digits;
     // do we add or subtract from start to reach end?
     bool forwards = end - start >= 0 ? true : false;
-    for(int i = start;
+    for(size_t i = start;
         forwards ? i <= end : i >= end;
         forwards ?      ++i :      --i )
     {
@@ -84,7 +93,7 @@ bool is_pandigital_for_range(int start, int end, size_t* ints, int ints_size)
     }
 
     bool keep_going = true;
-    for(int i = 0; i < ints_size; ++i)
+    for(size_t i = 0; i < ints_size; ++i)
     {
         keep_going = spy_the_digits(forwards ? start : end,
                                     forwards ? end : start,
@@ -107,7 +116,7 @@ bool is_pandigital_for_range(int start, int end, size_t* ints, int ints_size)
 
     // we did!
     std::cout << "([" << start << ".." << end << "], ";
-    for(int i=0;i<ints_size;++i)
+    for(size_t i=0;i<ints_size;++i)
     {
         std::cout << ints[i] << ",";
     }
@@ -115,7 +124,7 @@ bool is_pandigital_for_range(int start, int end, size_t* ints, int ints_size)
     return true;
 }
 
-bool spy_the_digits(int min, int max, size_t n, digit_map& reqd_digits)
+bool spy_the_digits(size_t min, size_t max, size_t n, digit_map& reqd_digits)
 {
     size_t left = n;
     while(left>0)
@@ -138,9 +147,9 @@ bool spy_the_digits(int min, int max, size_t n, digit_map& reqd_digits)
     return true;
 }
 
-int len_n(size_t n)
+size_t len_n(size_t n)
 {
-    int len = 1;
+    size_t len = 1;
     while(n>9)
     {
         ++len;
